requirements.txt
commentjson==0.9.0
cyclonedds==0.10.5
docutils
ghapi==1.0.5
imageio==2.26.1
nox==2024.4.15
numpy
opencv_python==4.6.0.66
Pillow==10.4.0
pyquaternion==0.9.9
pytest==8.3.2
pytorch3d
Requests==2.32.3
rich==13.7.1
scipy==1.9.1
setuptools
sympy==1.11.1
tensorflow
tensorflow_probability
torchviz==0.0.2
tqdm==4.65.0
trimesh==3.21.0
tqdm
numpy
trimesh